requirements.txt
python-telegram-bot==13.7
pymongo[srv]==4.3.3
python-dotenv==0.19.0
configparser